# backend/openapi.yaml
...
components:
  schemas:
...
    AffiliateHour:
      properties:
        end:
          type: string
        label:
          type: string
        start:
          type: string
      type: object
    AffiliateResponse:
      properties:
        affiliate:
          properties:
            carrd_link:
              type: string
            created_at:
              type: string
            details:
              type: string
            discord_link:
              type: string
            embed_color:
              type: string
            hours:
              items:
                properties:
                  end:
                    type: string
                  label:
                    type: string
                  start:
                    type: string
                type: object
              type: array
            id:
              format: int64
              type: integer
            location:
              type: string
            logo:
              type: string
            name:
              type: string
            owners:
              items:
                type: string
              type: array
            screenshot:
              type: string
            sort_order:
              type: integer
            subtitle:
              type: string
            timezone:
              type: string
          type: object
      type: object
...
    AffiliatesResponse:
      properties:
        affiliates:
          items:
            properties:
              carrd_link:
                type: string
              created_at:
                type: string
              details:
                type: string
              discord_link:
                type: string
              embed_color:
                type: string
              hours:
                items:
                  properties:
                    end:
                      type: string
                    label:
                      type: string
                    start:
                      type: string
                  type: object
                type: array
              id:
                format: int64
                type: integer
              location:
                type: string
              logo:
                type: string
              name:
                type: string
              owners:
                items:
                  type: string
                type: array
              screenshot:
                type: string
              sort_order:
                type: integer
              subtitle:
                type: string
              timezone:
                type: string
            type: object
          type: array
        webhook_url:
          type: string
      type: object
...
    Error:
      properties:
        error:
          type: string
      type: object
...
    OKResponse:
      properties:
        ok:
          type: boolean
      type: object
...
  securitySchemes:
    bearerAuth:
      description: Personal access token (pat_...). Also accepted as ?token= on the
        WebSocket upgrade.
      scheme: bearer
      type: http
    cookieAuth:
      description: SCS session cookie set by POST /api/auth (login).
      in: cookie
      name: session
      type: apiKey
info:
  description: |-
    JSON / WebSocket API for the **Senpan App Suite** (real-time Bingo, raffles, a
    festival lottery drum, Discord announcements, book-club reading lists, font/image
    hosting, and a themeable, permissioned admin dashboard).

    ## Conventions

    - All responses are `application/json` with `Cache-Control: no-store`.
    - Request bodies are JSON, capped at 1 MB (multipart uploads have their own caps).
    - Errors use the envelope `{ "error": "message" }` (the **Error** schema) with an
      appropriate status. A malformed JSON body returns `400 Invalid JSON`.
    - Most `POST` endpoints are **action dispatchers**: the body carries an `action`
      field and the server switches on it. Each operation lists its actions and the
      fields each one uses.

    ## Authentication

    Identity is resolved from either a **session cookie** (set by `POST /api/auth`
    login) or a **personal access token** (PAT) supplied as `Authorization: Bearer pat_...`
    (or `?token=pat_...` on the WebSocket upgrade). A PAT inherits its owning account's
    permissions, so external clients use the same guards as the SPA.

    Guard levels referenced in each operation's description:
    - **public** - no auth.
    - **auth** - any authenticated, active account.
    - **admin** - an active admin account.
    - **permission:<key>** - an admin, or a non-admin holding that page-permission key.

...
    (not a grantable key).

    ## WebSocket - `GET /api/ws`

...

...

...

...
openapi: 3.0.3
paths:
...
  /api/affiliates:
    get:
      description: |-
        **Auth:** permission:teahouse-affiliates

        Also returns the shared Discord webhook (`webhook_url`); safe here since the endpoint is permission-gated.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AffiliatesResponse'
          description: Success
      security:
      - cookieAuth: []
      - bearerAuth: []
      summary: List affiliates
      tags:
      - Affiliates
    post:
      description: '**Auth:** permission:teahouse-affiliates'
      requestBody:
        content:
          application/json:
            schema:
              properties:
                carrd_link:
                  description: Carrd (or other site) link.
                  type: string
                details:
                  description: Markdown.
                  type: string
                discord_link:
                  description: Discord link.
                  type: string
                embed_color:
                  description: Discord embed accent '#rrggbb'.
                  type: string
                hours:
                  items:
                    $ref: '#/components/schemas/AffiliateHour'
                  type: array
                location:
                  type: string
                logo:
                  type: string
                name:
                  description: Name (required).
                  type: string
                owners:
                  items:
                    type: string
                  type: array
                screenshot:
                  type: string
                subtitle:
                  description: Optional second line under the name (any script).
                  type: string
                timezone:
                  description: IANA zone.
                  type: string
              type: object
        description: Affiliate fields.
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AffiliateResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Name required
      security:
      - cookieAuth: []
      - bearerAuth: []
      summary: Create an affiliate
      tags:
      - Affiliates
  /api/affiliates/{id}:
    delete:
      description: '**Auth:** permission:teahouse-affiliates'
      parameters:
      - description: Affiliate id.
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
      security:
      - cookieAuth: []
      - bearerAuth: []
      summary: Delete an affiliate
      tags:
      - Affiliates
    put:
      description: |-
        **Auth:** permission:teahouse-affiliates

        Full replace of the editable fields (sort_order is preserved - reordering is separate).
      parameters:
      - description: Affiliate id.
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                carrd_link:
                  description: Carrd (or other site) link.
                  type: string
                details:
                  description: Markdown.
                  type: string
                discord_link:
                  description: Discord link.
                  type: string
                embed_color:
                  description: Discord embed accent '#rrggbb'.
                  type: string
                hours:
                  items:
                    $ref: '#/components/schemas/AffiliateHour'
                  type: array
                location:
                  type: string
                logo:
                  type: string
                name:
                  description: Name (required).
                  type: string
                owners:
                  items:
                    type: string
                  type: array
                screenshot:
                  type: string
                subtitle:
                  description: Optional second line under the name (any script).
                  type: string
                timezone:
                  description: IANA zone.
                  type: string
              type: object
        description: Full affiliate fields.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OKResponse'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Name required
      security:
      - cookieAuth: []
      - bearerAuth: []
      summary: Replace an affiliate
      tags:
      - Affiliates
...
servers:
...